+++ src/nativelib/__init__.py
@@ -33,4 +33,4 @@ __all__ = (
     "Size",
 )
 __author__ = "0xMihalich"
-__version__ = "0.2.4.dev0"
+__version__ = "0.2.4.dev1"